refactor(server): extract authorization state machine
